Renamed and moved Create Character DTO to Create Character Request.  Also changed it to a record
