+++ ExpressedRealms.Server/EndPoints/CharacterEndPoints/Requests/CreateCharacterRequestValidator.cs
@@ -1,13 +1,12 @@
 using ExpressedRealms.DB;
 using FluentValidation;
-using FluentValidation.Results;
 using Microsoft.EntityFrameworkCore;
 
-namespace ExpressedRealms.Server.EndPoints.DTOs;
+namespace ExpressedRealms.Server.EndPoints.CharacterEndPoints.Requests;
 
-public class CreateCharacterDTOValidator : AbstractValidator<CreateCharacterDTO>
+public class CreateCharacterRequestValidator : AbstractValidator<CreateCharacterRequest>
 {
-    public CreateCharacterDTOValidator(ExpressedRealmsDbContext dbContext)
+    public CreateCharacterRequestValidator(ExpressedRealmsDbContext dbContext)
     {
         RuleFor(x => x.Name).NotEmpty().MaximumLength(150);
         RuleFor(x => x.ExpressionId).NotEmpty();
@@ -18,7 +17,8 @@ public class CreateCharacterDTOValidator : AbstractValidator<CreateCharacterDTO>
             RuleFor(x => x.ExpressionId).MustAsync(async (expressionId, cancellationToken) =>
             {
                 return await dbContext.Expressions.AnyAsync(x => x.Id == expressionId, cancellationToken);
-            }).WithMessage("This is not a valid Expression Id");            
+            })
+                .WithMessage("This is not a valid Expression Id");            
             
             RuleFor(x => x).MustAsync(async (dto, cancellationToken) =>
             {
@@ -26,7 +26,9 @@ public class CreateCharacterDTOValidator : AbstractValidator<CreateCharacterDTO>
                     .AnyAsync(x => x.ExpressionId == dto.ExpressionId 
                                    && x.SectionTypeId == (int)ExpressionSectionType.FactionType
                                    && x.Id == dto.FactionId, cancellationToken);
-            }).WithName(nameof(CreateCharacterDTO.FactionId)).WithMessage("This is not a valid Faction Id");
+            })
+                .WithName(nameof(CreateCharacterRequest.FactionId))
+                .WithMessage("This is not a valid Faction Id");
         });
     }
 }
